% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{add_logo}
\alias{add_logo}
\alias{add_copyright}
\alias{add_watermark}
\title{Add Marks}
\usage{
add_logo(ggobj, logo.size = 0.1, logo.loc = c(0.5, 0.5))

add_copyright(ggobj, just = c("center", "bottom"),
  text = "copyright: www.yikdata.com", col = "black")

add_watermark(ggobj, text = "www.yikdata.com", angle = 0,
  alpha = 0.4)
}
\arguments{
\item{ggobj}{A ggplot object.}

\item{logo.size}{Controls the size of logo image.}

\item{logo.loc}{Coordinates of the center point, which controls the position of logo image.}

\item{just}{A string or numeric
    vector specifying the justification of the viewport
    relative to its (x, y) location.  If there are two values, the first
    value specifies horizontal justification and the second value specifies
    vertical justification.  Possible string values are: \code{"left"},
    \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"},
    and \code{"top"}.  For numeric values, 0 means left alignment
    and 1 means right alignment.
  }

\item{text}{Character to add to the figure.}

\item{col}{Colour for text.}

\item{angle}{The angle of text.}

\item{alpha}{Alpha channel for transparency}
}
\description{
Include logo, copyright statement and watermarks.
}
