% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_cont.R
\name{fa_business_scale}
\alias{fa_business_scale}
\title{Closure function to create continuous scales.}
\usage{
fa_business_scale(aesthetics = "fill")
}
\description{
Closure function to create continuous scales.
}
