% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_bubble_rect.R
\name{fig_bubble_rect}
\alias{fig_bubble_rect}
\title{Bubble Graph}
\usage{
fig_bubble_rect(rawdata, text.size.x = 12, text.size.y = 12,
  palette = "Spectral", bubble.size.range = c(3, 15))
}
\arguments{
\item{rawdata}{A \code{data.frame} with three columns as follows:
'xlabel',character vector for discrete variables in x axis;
'ylabel',character vector for discrete variables in y axis;
'values',numeric vector defines the data value to be ploted.}

\item{text.size.x}{The text size of x axis.}

\item{text.size.y}{The text size of y axis.}

\item{palette}{The palette used for points.}

\item{bubble.size.range}{A \code{vector} with length two, defines the size of
points, the format is like c(minsize, maxsize).}
}
\description{
This function gives a way to draw a point matrix, with both the x and y axises
are discrete variables and map the values to the size of points.
}
\examples{
numx <- 10
numy <- 7
rawdata <- data.frame(
  xlabel = rep(paste(letters[1:numx], 'x', sep = '-'), numy),
  ylabel = rep(paste(letters[1:numy], 'y', sep = '-'), each = numx),
  values = runif(numx*numy, min = 5, max = 30)
)

fig_bubble_rect(rawdata)
}
\seealso{
Other figure: \code{\link{fig_calendar}},
  \code{\link{fig_event_storm}}
}
\concept{figure}
