% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_calendar.R
\name{fig_calendar}
\alias{fig_calendar}
\title{Calendar chart}
\usage{
fig_calendar(rawdata, colours = c("beige", "orange", "red", "darkred"),
  alpha = 0.8, na.col = "grey80", val.col = "grey70",
  mar.col = "grey80", mar.fill = "white", strip.text.size = 12,
  strip.text.hjust = 0, axis.x.size = 10, isfull = F, islegend = T,
  istext = F, text.size = 3, isday = F, day.size = 3)
}
\arguments{
\item{rawdata}{A \code{data.frame} with two columns: 'dates', date data in the
format of \code{Date()}; 'values', numeric data.}

\item{colours}{A \code{vector} contains colours mapped to the figure.}

\item{alpha}{Numeric value between 1 and 0, controls the transparency of rects.}

\item{na.col}{Colour of the NA values.}

\item{val.col}{Border colour of the rects.}

\item{mar.col}{Border colour of the rects, which used to fill blank space.}

\item{mar.fill}{Colour used to fill the rects, which used to fill black space.}

\item{strip.text.size}{Text size of strip.}

\item{strip.text.hjust}{Horizontal position of the text in strip.}

\item{axis.x.size}{Text size of axis x.}

\item{isfull}{Boolean value, default FALSE,if TRUE use rects fill the blank space.}

\item{islegend}{Boolean value, default TRUE, if FALSE remove the legend.}

\item{istext}{Boolean value, default False, if TRUE draw the value to the correspond date.}

\item{text.size}{Text size of data value.}

\item{isday}{Boolean value, default FALSE, if TRUE draw the date number.}

\item{day.size}{Text size of date number.}
}
\value{
ggplot object
}
\description{
Plot time series data in calendar way.
}
\note{
\code{istext} and \code{isday} can only set one to be TRUE at one time.
}
\examples{

# 1 month
rawdata <- data.frame(
  dates = seq(as.Date('2018-01-01'), length.out = 20, by = '1 day'),
  values = rnorm(20)
)
# fig_calendar(
#   rawdata,
#   na.col = 'grey90', val.col = 'grey80',
#   mar.fill = 'white', mar.col = 'grey80',
#   strip.text.size = 18, strip.text.hjust = 0.5,
#   axis.x.size = 12, isfull = TRUE, islegend = FALSE
# )

# 6 months
num <- 180
# rawdata <- data.frame(
#   dates = seq(as.Date('2018-01-01'), length.out = num, by = '1 day'),
#   values = round(rnorm(num), 3)
# )
# fig_calendar(rawdata, isfull = TRUE, na.col = 'grey80', val.col = 'grey80',
#              mar.fill = 'white', mar.col = 'grey80')

# add value tag
# fig_calendar(
#   rawdata,
#   na.col = 'grey90', val.col = 'grey80',
#   mar.fill = 'white', mar.col = 'grey80',
#   strip.text.size = 18, strip.text.hjust = 0.5,
#   axis.x.size = 12, isfull = TRUE, islegend = FALSE,
#   istext = TRUE
# )
#

}
\seealso{
Other figure: \code{\link{fig_bubble_rect}},
  \code{\link{fig_event_line}},
  \code{\link{fig_event_storm}}, \code{\link{fig_parallel}}
}
\concept{figure}
