% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_event_line.R
\name{fig_event_line}
\alias{fig_event_line}
\title{Event Line}
\usage{
fig_event_line(rawdata)
}
\description{
Event Line
}
\seealso{
Other figure: \code{\link{fig_bubble_rect}},
  \code{\link{fig_calendar}},
  \code{\link{fig_event_storm}}, \code{\link{fig_parallel}}
}
\concept{figure}
