% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_parallel.R
\name{fig_parallel}
\alias{fig_parallel}
\title{Parallel Chart}
\usage{
fig_parallel(rawdata, alpha = 0.8, add.point = FALSE,
  legend.size = 8, line.size = 0.5)
}
\arguments{
\item{rawdata}{A data.frame with data in long format, there are four columns:
'id', integer, identify each sample; 'group', character,group of each sample;
'variable', character, the parallel axises; 'value', numeric, values mapped to
each axis.}

\item{alpha}{Alpha channel for transparency.}

\item{add.point}{Whether add points to axises or not.}

\item{legend.size}{Size of legend key.}

\item{line.size}{Size of line.}
}
\description{
Parallel Chart
}
\examples{
library(reshape2)
rawdata <- iris
rawdata$id <- 1:nrow(iris)
rawdata <- melt(rawdata, id = c('id','Species'))

names(rawdata) <- c('id', 'group', 'variable', 'value')

fig_parallel(rawdata)

fig_parallel(rawdata, alpha = 0.3, add.point = TRUE)

fig_parallel(rawdata, alpha = 0.3, add.point = TRUE) +
  coord_polar()
}
\seealso{
Other figure: \code{\link{fig_bubble_rect}},
  \code{\link{fig_calendar}}, \code{\link{fig_event_line}},
  \code{\link{fig_event_storm}}
}
\concept{figure}
