% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_waterfall.R
\name{fig_waterfull}
\alias{fig_waterfull}
\title{Customizated figure of waterfull.}
\usage{
fig_waterfull(values, start_label = "start", end_label = "end")
}
\arguments{
\item{overlap}{\code{logical} Use the scale for overlapping points?}
}
\description{
waterfull with bars.
}
\note{
This function is customizated by using the \code{waterfall()} function
from package waterfalls
}
\examples{
library(ggplot2)
library(waterfalls)

fig_waterfull(values = c(200, -50, 70),
              start_label = '开始', end_label = '结束')

fig_waterfull(values = c(200, -50, -19, 70),
              start_label = '开始', end_label = '结束')

}
\references{
CRAN \url{https://cran.r-project.org/} waterfalls
}
\concept{figures}
