% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medical_cpi}
\alias{medical_cpi}
\title{CPI data of prices for medical treatment}
\format{A data frame with 376 rows and 6 variables:
\describe{
\item{time}{Time stamp of int}
\item{medi_type}{Way of treatment, with two types: '17', in clinic; 'q8',
in hospital}
\item{cpi}{Value of cpi(>1)}
\item{drug}{The drug factor which influence the value of cpi, in comparison
with material and treatment}
\item{material}{The material factor which influence the value of cpi.}
\item{treatment}{The treatment factor which influence the value of cpi.}
}}
\usage{
medical_cpi
}
\description{
A dataset containing the CPI value of medical treatment prices and other
attributes.
}
\keyword{datasets}
