% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_cont.R
\name{pal_red_cont}
\alias{pal_red_cont}
\alias{pal_blue_cont}
\title{Color palettes (continuous)}
\usage{
pal_red_cont(isseq = TRUE)

pal_blue_cont(isseq = TRUE)
}
\arguments{
\item{isseq}{\code{logical}, if TRUE called with a numeric vector with values
between 0 and 1 returns the corresponding values in the range the scale
maps to, if FALSE called with a integer value controls the number of colours
to be returned.}
}
\description{
Color palettes (continuous)
}
\examples{
library(scales)

# get colours by location
pal <- pal_red_cont()
cols <- pal(c(0.1, 0.3, 0.5, 0.7, 0.9))
plot(1, 1, xlim = c(0, 6), ylim = c(0, 6), type = "n", axes = FALSE,
     bty = "n", xlab = "", ylab = "")
rect(xleft = 0:4, ybottom = 3, xright = 1:5, ytop = 4, col = cols,
     border = "light grey")

# get colours by number
pal <- pal_red_cont(isseq = FALSE)
cols <- pal(5)
plot(1, 1, xlim = c(0, 6), ylim = c(0, 6), type = "n", axes = FALSE,
     bty = "n", xlab = "", ylab = "")
rect(xleft = 0:4, ybottom = 3, xright = 1:5, ytop = 4, col = cols,
     border = "light grey")
}
\concept{colour palette}
