% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_disc.R
\name{pal_business}
\alias{pal_business}
\alias{pal_academic}
\title{YIK color palette (discrete)}
\usage{
pal_business(isdark = FALSE)

pal_academic(isdark = FALSE)
}
\arguments{
\item{isdark}{Use the dark style colors.}
}
\description{
YIK color palette (discrete)
}
\examples{

}
\seealso{
Other colour business: \code{\link{scale_colour_business_light}}
}
\concept{colour business}
