% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_cont.R
\name{scale_fill_continuous_yk}
\alias{scale_fill_continuous_yk}
\alias{scale_colour_continuous_yk}
\alias{scale_color_continuous_yk}
\title{Scalses for continuous values}
\usage{
scale_fill_continuous_yk(..., col.scheme = "red", na.value = "grey50",
  guide = "colourbar")

scale_colour_continuous_yk(..., col.scheme = "red",
  na.value = "grey50", guide = "colourbar")

scale_color_continuous_yk(..., col.scheme = "red", na.value = "grey50",
  guide = "colourbar")
}
\arguments{
\item{col.sacheme}{Only valid with continuous scales, for now only two colour
schemes(\code{c('red', 'blue')}) are provided.}
}
\description{
Scalses for continuous values
}
