% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalses.R
\name{scl_circle}
\alias{scl_circle}
\title{Scale circle}
\usage{
scl_circle(vdata, break_angles = 100)
}
\arguments{
\item{vdata}{A \code{vector} variable contains values to be scaled.}

\item{break_angles}{A \code{integer} value, controls the points to be generate.

family scale

export}
}
\description{
Scale circle
}
