% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_report.R
\name{theme_rp_light}
\alias{theme_rp_light}
\title{Theme for reports}
\usage{
theme_rp_light(font.size.legend.title = 12, font.size.legend.text = 10,
  font.size.axis.x = 12, font.size.axis.y = 12)
}
\arguments{
\item{font.size.legend.title}{Font size of legend title.}

\item{font.size.legend.text}{Font size of legend text.}

\item{font.size.axis.x}{Font size of axis x.}

\item{font.size.axis.y}{Font size of axis y.}
}
\description{
This theme is used for create figures in data analysis reports.
}
\examples{
library("ggplot2")

# Line and scatter plot colors
p <- ggplot(mtcars) +
  geom_point(aes(x = wt, y = mpg, colour = factor(gear))) +
  facet_wrap(~am)
p + theme_rp_light()

}
\seealso{
Other themes: \code{\link{theme_yk_academic}},
  \code{\link{theme_yk_business_brief}}
}
\concept{themes}
