% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R
\name{theme_yk_academic_light}
\alias{theme_yk_academic_light}
\alias{theme_yk_academic_dark}
\alias{theme_yk_academic_dotted_gray}
\alias{theme_yk_academic_brief}
\alias{theme_yk_academic_common}
\title{ggplot color theme for academic files}
\usage{
theme_yk_academic_light(base_size = 10, base_family = "")

theme_yk_academic_dark(base_size = 10, base_family = "")

theme_yk_academic_dotted_gray(base_size = 10, base_family = "")

theme_yk_academic_brief(base_size = 10, base_family = "")

theme_yk_academic_common(base_size = 10, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A theme that can be used in \emph{academic} scenes.
}
\details{
Use \code{\link{scale_color_academic_light}()} with this theme.
}
\examples{
data(medical_cpi)
head(medical_cpi)

medical_cpi$time <- as.Date(paste0(medical_cpi$time, '01'), format = '\%Y\%m\%d')

windowsFonts(yahei = windowsFont("微软雅黑"))
p <- ggplot(medical_cpi) +
  geom_line(aes(x = time, y = cpi, col = medi_type), size = 1)

# academic light
p + theme_yk_academic_light(base_family = 'yahei') +
  scale_color_academic_light()

# academic dark
p + theme_yk_academic_dark(base_family = 'yahei') +
  scale_color_academic_dark()
}
\seealso{
Other themes: \code{\link{theme_rp_light}},
  \code{\link{theme_yk_business_brief}}
}
\concept{themes}
