% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_business.R
\name{theme_yk_business_brief}
\alias{theme_yk_business_brief}
\alias{theme_yk_business_light}
\alias{theme_yk_business_dark}
\alias{theme_yk_business_light_solid}
\alias{theme_yk_business_dotted_lightblue}
\alias{theme_yk_business_blue}
\title{ggplot color theme for bussiness files}
\usage{
theme_yk_business_brief(base_size = 8, base_family = "")

theme_yk_business_light(base_size = 10, base_family = "")

theme_yk_business_dark(base_size = 10, base_family = "")

theme_yk_business_light_solid(base_size = 10, base_family = "")

theme_yk_business_dotted_lightblue(base_size = 10, base_family = "")

theme_yk_business_blue(base_size = 10, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A theme that can be used in \emph{bussiness} scenes.
}
\details{
Use \code{\link{scale_color_business_light}()} with this theme.
}
\examples{
data(medical_cpi)
head(medical_cpi)

medical_cpi$time <- as.Date(paste0(medical_cpi$time, '01'), format = '\%Y\%m\%d')

windowsFonts(yahei = windowsFont("微软雅黑"))
p <- ggplot(medical_cpi) +
  geom_line(aes(x = time, y = cpi, col = medi_type), size = 1)

# light
p + theme_yk_business_light(base_family = 'yahei') +
  scale_color_business_light()

# dark
p + theme_yk_business_dark(base_family = 'yahei') +
  scale_color_business_dark()

# continuous scale
p2 <- ggplot(medical_cpi) +
  geom_point(aes(x = time, y = cpi, colour = cpi, fill = cpi), size = 5, shape = 21)

p2 + theme_yk_business_light(base_family = 'yahei') +
  scale_colour_business_continuous(col.scheme = 'blue') +
  scale_fill_business_continuous(col.scheme = 'red')
}
\seealso{
Other themes: \code{\link{theme_rp_light}},
  \code{\link{theme_yk_academic}}
}
\concept{themes}
