% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_halfcircles.R
\name{fig_halfcircles}
\alias{fig_halfcircles}
\alias{fig_halfcircles_emph}
\title{Circle Chart for Data Comparison}
\usage{
fig_halfcircles(rawdata, title, subtitle, caption, palette = "Oranges")

fig_halfcircles_emph(rawdata, title, subtitle, caption,
  colours = c("white", "orange"))
}
\arguments{
\item{rawdata}{A \code{data.frame} contains the data to be plot. Three columns
are needed here, 'values': numeric, provide datas; 'types': used to map colours
to each data, character or factor; 'labels': characters to describe the correspond
data.}

\item{title}{\code{character} value defines the title.}

\item{subtitle}{\code{character} value defines the subtitle.}

\item{caption}{\code{character} value defines the caption.}

\item{palette}{Palette applied to the plot.}

\item{colours}{A \code{vector} of length two, contains colours which applied
to different data.}
}
\description{
Show all datas in a sequential way, and emphasis the special one.
}
\examples{
rawdata <- data.frame(
  values = c(220, 313, 393.4, 423.4, 583.4, 623.4, 823.4,
             934.5, 985.6),
  types = letters[1:9],
  labels = c(
    '原型构建', '参数调试', '模型优化', '模型训练', '业务理解',
    '资料检索', '数据采集(爬虫、文档、线上、线下)',
    '分析系统开发(UI设计, 数据库设计, 模型部署, 模型应用)', '数据清洗'),
  stringsAsFactors = F
)
fig_halfcircles(rawdata, title = '数据分析各环节耗时', subtitle = 'hours',
                caption = '注：数据纯属虚构')


rawdata <- data.frame(
  values = c(220, 313, 393.4, 423.4, 583.4, 623.4, 823.4,
             934.5, 985.6),
  types = c('a','a','a','a','a','a', 'b', 'a', 'a'),
  labels = c(
    '原型构建', '参数调试', '模型优化', '模型训练', '业务理解',
    '资料检索', '数据采集(爬虫、文档、线上、线下)',
    '分析系统开发(UI设计, 数据库设计, 模型部署, 模型应用)', '数据清洗'),
  stringsAsFactors = F
)
fig_halfcircles_emph(rawdata, title = '数据分析各环节耗时', subtitle = 'hours',
                     caption = '注：数据纯属虚构')

}
\seealso{
\code{\link{fig_halfcircles_emph}()} for references.

\code{\link{fig_halfcircles}()} for references.
}
\concept{figure circle}
