% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_waterfall.R
\name{fig_waterfall}
\alias{fig_waterfall}
\title{Customizated figure of waterfull.}
\usage{
fig_waterfall(values, colors = c(start = "#558ED5", increase = "#51A047",
  reduce = "#F1B255"), start_label = "start", end_label = "end")
}
\arguments{
\item{values}{Vector contains increase or reduce data.}

\item{colors}{Named vector, defines the colours used to fill the rects.}

\item{start_label}{Character.}

\item{end_label}{Character}
}
\description{
waterfull with bars.
}
\note{
This function is customizated by using the \code{waterfall()} function
from package waterfalls
}
\examples{
fig_waterfall(values = c(200, -50, 70),
              start_label = 'start', end_label = 'end')

fig_waterfall(values = c(200, -50, -19, 70),
              start_label = 'start', end_label = 'end')

fig_waterfall(values = c(200, -50, -19, 70, 30, 13, -5, 19),
              start_label = 'start', end_label = 'end')

}
\references{
CRAN \url{https://cran.r-project.org/} waterfalls
}
\seealso{
Other figures: \code{\link{fig_gantt}}
}
\concept{figures}
