% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_disc.R
\name{yik_palette_disc}
\alias{yik_palette_disc}
\alias{pal_business_light}
\alias{pal_business_dark}
\alias{pal_business_light2}
\alias{pal_business_dark2}
\alias{pal_business_bright}
\alias{pal_business_grey}
\alias{pal_business_warm}
\alias{pal_business_cool}
\alias{pal_academic_light}
\alias{pal_academic_dark}
\title{YIK color palette (discrete)}
\usage{
pal_business_light(isdark = FALSE)

pal_business_dark(isdark = FALSE)

pal_business_light2(isdark = FALSE)

pal_business_dark2(isdark = FALSE)

pal_business_bright()

pal_business_grey()

pal_business_warm()

pal_business_cool()

pal_academic_light()

pal_academic_dark(isdark = FALSE)
}
\arguments{
\item{isdark}{Use the dark style colors.}
}
\description{
YIK color palette (discrete)
}
\examples{

}
\seealso{
Other colour business: \code{\link{scale_colour_business_light}}
}
\concept{colour business}
